Add files via upload
The Forza Horizon Cars Dataset is a comprehensive collection of data on cars featured in the Forza Horizon series, including details on their attributes, availability, and acquisition methods. This dataset is utilized for analysis and visualization purposes, particularly in Power BI, a powerful business intelligence tool. By leveraging Power BI, users can conduct in-depth data analysis, create interactive visualizations, and gain valuable insights into the characteristics and distribution of cars within the Forza Horizon game series. The dataset serves as a valuable resource for enthusiasts and researchers interested in exploring the diverse range of cars available in the game, tracking their progress, and optimizing gameplay strategies.